requirements.txt
Flask>=2.2
Flask-Session>=0.4
Authlib>=1.2
SQLAlchemy>=1.4
requests>=2.28
python-dotenv>=1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pdfplumber>=0.7.8
nltk>=3.8.1
werkzeug>=2.2
gunicorn
